edge_node_pi/requirement.txt
opencv-python
pyyaml
onnxruntime
numpy
RPi.GPIO
ultralytics